Ignore write errors when forcing a refresh
Fixes #32.

On the production server, the scraping cronjob runs as ludumdare but the
server runs as www-data. Consequently, the "Force refresh" option cannot
overwrite images written by the scraper.

A better fix would be to change the group ownership of the directory to
be www-data but keep the owner as ludumdare. Then, we first write the
file to a temporary name in the same directory, then atomically move it
into the place of the old one.

But I have neither time nor courage to do either right now, so this
workaround will have to suffice.